rally-point: add per-building rally destinations with UnitProductionSystem wiring
RallyPointComponent (per-building) and PlayerRallyPointSetRequestComponent (one-shot)
define the rally state; RallyPointSetSystem processes set/clear requests.
UnitProductionSystem reads RallyPointComponent at spawn time and routes each new unit's
MoveCommandComponent to the rally position when active. Debug surface, 3-phase smoke
validator, PS1 wrapper, and csproj entries included. Contract bumped to revision 122.
Also fixes CS0234 (Unity.Collections.ComponentType) and CS0121 (math.min byte ambiguity)
in siege-escalation and dynasty-progression lane files caught at build gate.

+++ bloodlines/unity/Assets/_Bloodlines/Code/Dynasties/DynastyProgressionCanon.cs
@@ -46,7 +46,7 @@ namespace Bloodlines.Dynasties
                 }
             }
 
-            return math.min(tier, MaxTier);
+            return (byte)math.min((int)tier, (int)MaxTier);
         }
 
         /// Returns the XP threshold to reach the next tier, or float.MaxValue if at cap.
